fix: fix duration formatting
#1
